Added scaling factors to fits
Not all measured dose-response curves reach minimum FRET. This is likely because only one ligand-species is not sufficient to fully inactivate the arrays. These edits add in scaling factors that scale the fits of the CDF and hill function to this kind of data. Seems at a glance to improve fitting quality.